Move IFactory to shape dir
