+++ src/Shapes/IFactory.ts
